+++ src/test/java/org/opensearch/knn/index/SpaceTypeTests.java
@@ -16,16 +16,19 @@ import org.opensearch.knn.KNNTestCase;
 import org.opensearch.knn.index.util.KNNEngine;
 
 import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
+import java.util.Map;
+import java.util.Set;
 
 public class SpaceTypeTests extends KNNTestCase {
 
     public void testGetVectorSimilarityFunction_l2() {
-        assertEquals(VectorSimilarityFunction.EUCLIDEAN, SpaceType.L2.getVectorSimilarityFunction());
+        assertEquals(VectorSimilarityFunction.EUCLIDEAN, SpaceType.L2.getKnnVectorSimilarityFunction().getVectorSimilarityFunction());
     }
 
     public void testGetVectorSimilarityFunction_invalid() {
-        expectThrows(UnsupportedOperationException.class, SpaceType.L1::getVectorSimilarityFunction);
+        expectThrows(UnsupportedOperationException.class, SpaceType.L1::getKnnVectorSimilarityFunction);
     }
 
     public void testGetVectorSimilarityFunction_whenInnerproduct_thenConsistentWithScoreTranslation() {
@@ -56,9 +59,45 @@ public class SpaceTypeTests extends KNNTestCase {
         for (int i = 0; i < dataVectors.size(); i++) {
             assertEquals(
                 KNNEngine.FAISS.score(dotProducts.get(i), SpaceType.INNER_PRODUCT),
-                SpaceType.INNER_PRODUCT.getVectorSimilarityFunction().compare(queryVector, dataVectors.get(i)),
+                SpaceType.INNER_PRODUCT.getKnnVectorSimilarityFunction().compare(queryVector, dataVectors.get(i)),
                 0.0000001
             );
+        }
+    }
+
+    public void testValidateVectorDataType_whenCalled_thenReturn() {
+        Map<SpaceType, Set<VectorDataType>> expected = Map.of(
+            SpaceType.UNDEFINED,
+            Collections.emptySet(),
+            SpaceType.L2,
+            Set.of(VectorDataType.FLOAT, VectorDataType.BYTE),
+            SpaceType.COSINESIMIL,
+            Set.of(VectorDataType.FLOAT, VectorDataType.BYTE),
+            SpaceType.L1,
+            Set.of(VectorDataType.FLOAT, VectorDataType.BYTE),
+            SpaceType.LINF,
+            Set.of(VectorDataType.FLOAT, VectorDataType.BYTE),
+            SpaceType.INNER_PRODUCT,
+            Set.of(VectorDataType.FLOAT, VectorDataType.BYTE),
+            SpaceType.HAMMING_BIT,
+            Set.of(VectorDataType.BINARY)
+        );
+
+        for (SpaceType spaceType : SpaceType.values()) {
+            for (VectorDataType vectorDataType : VectorDataType.values()) {
+                if (expected.get(spaceType).isEmpty()) {
+                    Exception ex = expectThrows(IllegalStateException.class, () -> spaceType.validateVectorDataType(vectorDataType));
+                    assertTrue(ex.getMessage().contains("Unsupported method"));
+                    continue;
+                }
+
+                if (expected.get(spaceType).contains(vectorDataType)) {
+                    spaceType.validateVectorDataType(vectorDataType);
+                } else {
+                    Exception ex = expectThrows(IllegalArgumentException.class, () -> spaceType.validateVectorDataType(vectorDataType));
+                    assertTrue(ex.getMessage().contains("is not supported"));
+                }
+            }
         }
     }
 }