+++ src/test/java/org/opensearch/knn/index/query/filtered/FilteredIdsKNNByteIteratorTests.java
@@ -5,14 +5,13 @@
 
 package org.opensearch.knn.index.query.filtered;
 
+import junit.framework.TestCase;
 import lombok.SneakyThrows;
 import org.apache.lucene.index.BinaryDocValues;
 import org.apache.lucene.search.DocIdSetIterator;
 import org.apache.lucene.util.BytesRef;
 import org.apache.lucene.util.FixedBitSet;
-import org.opensearch.knn.KNNTestCase;
 import org.opensearch.knn.index.SpaceType;
-import org.opensearch.knn.index.codec.util.KNNVectorAsArraySerializer;
 
 import java.util.Arrays;
 import java.util.List;
@@ -21,25 +20,19 @@ import java.util.stream.Collectors;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-public class FilteredIdsKNNIteratorTests extends KNNTestCase {
+public class FilteredIdsKNNByteIteratorTests extends TestCase {
     @SneakyThrows
     public void testNextDoc_whenCalled_IterateAllDocs() {
-        final SpaceType spaceType = SpaceType.L2;
-        final float[] queryVector = { 1.0f, 2.0f, 3.0f };
+        final SpaceType spaceType = SpaceType.HAMMING_BIT;
+        final byte[] queryVector = { 1, 2, 3 };
         final int[] filterIds = { 1, 2, 3 };
-        final List<float[]> dataVectors = Arrays.asList(
-            new float[] { 11.0f, 12.0f, 13.0f },
-            new float[] { 14.0f, 15.0f, 16.0f },
-            new float[] { 17.0f, 18.0f, 19.0f }
-        );
+        final List<byte[]> dataVectors = Arrays.asList(new byte[] { 11, 12, 13 }, new byte[] { 14, 15, 16 }, new byte[] { 17, 18, 19 });
         final List<Float> expectedScores = dataVectors.stream()
             .map(vector -> spaceType.getKnnVectorSimilarityFunction().compare(queryVector, vector))
             .collect(Collectors.toList());
 
         BinaryDocValues values = mock(BinaryDocValues.class);
-        final List<BytesRef> byteRefs = dataVectors.stream()
-            .map(vector -> new BytesRef(new KNNVectorAsArraySerializer().floatToByteArray(vector)))
-            .collect(Collectors.toList());
+        final List<BytesRef> byteRefs = dataVectors.stream().map(vector -> new BytesRef(vector)).collect(Collectors.toList());
         when(values.binaryValue()).thenReturn(byteRefs.get(0), byteRefs.get(1), byteRefs.get(2));
 
         FixedBitSet filterBitSet = new FixedBitSet(4);
@@ -49,7 +42,7 @@ public class FilteredIdsKNNIteratorTests extends KNNTestCase {
         }
 
         // Execute and verify
-        FilteredIdsKNNIterator iterator = new FilteredIdsKNNIterator(filterBitSet, queryVector, values, spaceType);
+        FilteredIdsKNNByteIterator iterator = new FilteredIdsKNNByteIterator(filterBitSet, queryVector, values, spaceType);
         for (int i = 0; i < filterIds.length; i++) {
             assertEquals(filterIds[i], iterator.nextDoc());
             assertEquals(expectedScores.get(i), (Float) iterator.score());